requirements.txt
chainlit
openai
supabase
psycopg2-binary
python-dotenv
